Changed candidate to donor/ updated table fordonor
